feat(validation): rename, move and export getItem options interface
