+++ src/lib/index.ts
@@ -1,3 +1,3 @@
-export { AsyncLocalStorage } from './src/service/lib.service';
+export { JSONSchema, JSONSchemaType } from './src/service/validation/json-schema';
+export { AsyncLocalStorage, ALSGetItemOptions } from './src/service/lib.service';
 export { AsyncLocalStorageModule } from './src/module';
-export { JSONSchema, JSONSchemaType } from './src/service/validation/index';
