+++ src/lib/src/service/databases/async-local-database.ts
@@ -1,10 +1,6 @@
 import { Observable } from 'rxjs/Observable';
 
-import { JSONSchema } from '../validation/index';
-
-export interface GetItemOptions {
-  schema?: JSONSchema | null;
-}
+import { ALSGetItemOptions } from '../lib.service';
 
 export abstract class AsyncLocalDatabase {
 
@@ -12,7 +8,7 @@ export abstract class AsyncLocalDatabase {
     schema: null
   };
 
-  abstract getItem<T = any>(key: string, options?: GetItemOptions): Observable<T | null>;
+  abstract getItem<T = any>(key: string, options?: ALSGetItemOptions): Observable<T | null>;
   abstract setItem(key: string, data: any): Observable<boolean>;
   abstract removeItem(key: string): Observable<boolean>;
   abstract clear(): Observable<boolean>;
