+++ src/lib/src/service/databases/index.ts
@@ -1,4 +1,4 @@
-export { AsyncLocalDatabase, GetItemOptions } from './async-local-database';
+export { AsyncLocalDatabase } from './async-local-database';
 export { IndexedDBDatabase } from './indexeddb-database';
 export { LocalStorageDatabase } from './localstorage-database';
 export { MockLocalDatabase } from './mock-local-database';
