+++ src/lib/src/service/databases/mock-local-database.ts
@@ -3,8 +3,9 @@ import { Observable } from 'rxjs/Observable';
 import { map } from 'rxjs/operators';
 import { of as observableOf } from 'rxjs/observable/of';
 
-import { AsyncLocalDatabase, GetItemOptions } from './async-local-database';
-import { JSONValidator } from '../validation/index';
+import { AsyncLocalDatabase } from './async-local-database';
+import { ALSGetItemOptions } from '../lib.service';
+import { JSONValidator } from '../validation/json-validator';
 
 @Injectable()
 export class MockLocalDatabase extends AsyncLocalDatabase {
@@ -22,7 +23,7 @@ export class MockLocalDatabase extends AsyncLocalDatabase {
    * @param key The item's key
    * @returns The item's value if the key exists, null otherwise, wrapped in an RxJS Observable
    */
-   getItem<T = any>(key: string, options: GetItemOptions = this.getItemOptionsDefault) {
+   getItem<T = any>(key: string, options: ALSGetItemOptions = this.getItemOptionsDefault) {
 
     const rawData: T | null = this.localStorage.get(key);
